[CSM-33] Use mocks for integration tests
The tests here should likely be replicated with a full server install,
but for now, update them with mocks in order to allow tox to run
locally.

The original form of the tests have been moved over to
functional_tests.py, so we don't lose the full end-to-end coverage on
Travis jobs.