Use a bash script to boostrap mongodb
Abell expects an admin user to be preset when it does it's DB setup, but
MongoDB does not have one by default.  This script will use the MongoDB
localhost connection to bootstrap an admin user, both for local installs
(as opposed to Docker, which handles this via env vars) and CI systems.