Add basic client testing
This commit adds a scaffolding for testing the Abell client. Right now,
we're relying on unittest's mock objects to confirm that we're passing
on the correct arguments through to the requests.Session object.

The requests_mock library is added so that we can later mock the HTTP
API provided by Abell, without necessarily standing up a new instance
all the time.